+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 10.7
 - CSS add Select2 styles in stylesheet.css & colors.css
 - JS add Select2 jQuery plugin in assets/js/jquery-select2/
 - Add Select2Input() function in Inputs.php
+- Use Select2 instead of Chosen, fixes the overflow issue in Widget.php, MakeReferral.php, Transcripts.fnc.php & Address.inc.php
 
 Changes in 10.6.3
 -----------------